+++ exports/OwcTable.types.ts
@@ -3,6 +3,8 @@ export {
   handleUpdate,
 } from '../src/field-path-helper/getFieldPathContent.types.js';
 
+import { OwcTable } from './OwcTable.js';
+
 export {
   NestedJsonFilters,
   operator,
@@ -12,3 +14,9 @@ export {
 } from '../src/filter/filter.type.js';
 
 export { Column } from '../src/table/OwcTable.types.js';
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'owc-table': OwcTable<Record<string, unknown>>;
+  }
+}