+++ exports/OwcTableFilterBuilder.js
@@ -1,1 +1,2 @@
 export { OwcTableFilterBuilder } from '../src/table-filter/OwcTableFilterBuilder.js';
+export { OwcTableFilter } from '../src/table-filter/OwcTableFilter.js';