+++ exports/csv.js
@@ -1,1 +1,0 @@
-export { convertToCsv, copyAsCsv, downloadAsCsv } from '../src/table/csv.js';