+++ exports/excel.js
@@ -1,1 +1,0 @@
-export { convertToExcel, copyAsExcel } from '../src/table/excel.js';