+++ exports/filter.js
@@ -1,4 +1,0 @@
-export { globalSearchField, jsonToFilter } from '../src/filter/jsonToFilter.js';
-export { jsonToSqlFilter } from '../src/filter/jsonToSqlFilter.js';
-
-export { dateParserForJsonDecode } from '../src/table/dateParserForJsonDecode.js';