+++ exports/filter/jsonToFilter.js
@@ -1,0 +1,1 @@
+export { globalSearchField, jsonToFilter } from '../../src/filter/jsonToFilter.js';