+++ exports/filter/types.ts
@@ -1,0 +1,10 @@
+export {
+  Filter,
+  JsonFilter,
+  NestedJsonFilters,
+  SqlPrimitive,
+  dateOperator,
+  numberOperator,
+  operator,
+  textOperator,
+} from '../../src/filter/filter.type.js';