+++ exports/highlightSearchTerms.js
@@ -1,1 +1,0 @@
-export { highlightSearchTerms } from '../src/highlight-search-terms/highlightSearchTerms.js';