+++ exports/lit-helpers.js
@@ -1,1 +1,0 @@
-export { litHtmlToString } from '../src/lit-helpers/litHtmlToString.js';