+++ exports/lit/litHtmlToString.js
@@ -1,0 +1,1 @@
+export { litHtmlToString } from '../../src/lit-helpers/litHtmlToString.js';