+++ exports/text/highlightSearchTerms.js
@@ -1,0 +1,1 @@
+export { highlightSearchTerms } from '../../src/highlight-search-terms/highlightSearchTerms.js';